requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asttokens @ file:///C:/b/abs_9662ywy9fp/croot/asttokens_1743630464377/work
certifi==2025.8.3
charset-normalizer==3.4.3
colorama @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/colorama_1729036581634/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
distro==1.9.0
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
filelock==3.19.1
fsspec==2025.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
ipython @ file:///C:/b/abs_307wfft5yu/croot/ipython_1744912982439/work
ipython_pygments_lexers @ file:///C:/b/abs_b66hj1lo19/croot/ipython_pygments_lexers_1744753262043/work
jedi @ file:///C:/b/abs_3a2kbnlclc/croot/jedi_1733987412687/work
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline @ file:///C:/Users/dev-admin/buildout/croot/matplotlib-inline_1732914206474/work
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
openai==1.102.0
packaging==25.0
parso @ file:///C:/b/abs_834b4mj92b/croot/parso_1733963322289/work
pgvector==0.4.1
pillow==11.3.0
prompt_toolkit @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/prompt-toolkit_1729096615940/work
psycopg2-binary==2.9.10
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pydantic==2.11.7
pydantic_core==2.33.2
Pygments @ file:///C:/b/abs_e4bg5vh5j_/croot/pygments_1744667628203/work
PyMuPDF==1.26.4
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.8.29
requests==2.32.5
rich==14.1.0
ripgreppythonfiles==0.10
ruff==0.12.10
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
setuptools==78.1.1
sniffio==1.3.1
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.0
torch==2.8.0
touchtouch==0.11
tqdm==4.67.1
traitlets @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/traitlets_1729038154689/work
transformers==4.56.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth @ file:///C:/b/abs_0c8p4c33bp/croot/wcwidth_1750352902378/work
wheel==0.45.1
